refactor: convert `<NuxtImg>` to SFC
